refactor: share qwen3.5 mtp model logic across backends.
